+++ lib/mzgl/util/maths.h
@@ -13,8 +13,8 @@ float mapf(float inp, float inMin, float inMax, float outMin, float outMax, bool
 double mapd(double inp, double inMin, double inMax, double outMin, double outMax, bool clamp = false);
 
 
-float clampi(int inp, int from, int to);
-float clampl(long inp, long from, long to);
+int clampi(int inp, int from, int to);
+long clampl(long inp, long from, long to);
 float clampf(float inp, float from, float to);
 double clampd(double inp, double from, double to);
